docs: Separate `branding` and `theme` sections
