+++ docs/data/toolpad/core/components/sign-in-page/BrandingSignInPage.tsx
@@ -1,13 +1,9 @@
 import * as React from 'react';
 import { AuthProvider, AppProvider, SignInPage } from '@toolpad/core';
-import { createTheme } from '@mui/material/styles';
-import { useColorSchemeShim } from 'docs/src/modules/components/ThemeContext';
-import { getDesignTokens } from './brandingTheme';
 
 const providers = [
   { id: 'github', name: 'GitHub' },
   { id: 'google', name: 'Google' },
-  { id: 'credentials', name: 'Email and Password' },
 ];
 // preview-start
 const BRANDING = {
@@ -33,20 +29,9 @@ const signIn: (provider: AuthProvider) => void = async (provider) => {
 };
 
 export default function BrandingSignInPage() {
-  const { mode, systemMode } = useColorSchemeShim();
-  const calculatedMode = (mode === 'system' ? systemMode : mode) ?? 'light';
-  const brandingDesignTokens = getDesignTokens(calculatedMode);
-  const THEME = createTheme({
-    ...brandingDesignTokens,
-    palette: {
-      ...brandingDesignTokens.palette,
-      mode: calculatedMode,
-    },
-  });
-
   return (
     // preview-start
-    <AppProvider branding={BRANDING} theme={THEME}>
+    <AppProvider branding={BRANDING}>
       <SignInPage signIn={signIn} providers={providers} />
     </AppProvider>
     // preview-end
